app/requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.24
django-cors-headers==3.0.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.2
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
idna==2.8
isort==4.3.20
itypes==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
requests==2.25.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.26.5
wrapt==1.11.1
